+++ src/controllers/UtilityController.ts
@@ -1,8 +1,8 @@
 import { Body, Post, Queries, Route, Security, Tags } from "tsoa/dist";
 
 import { respondFailure, respondSuccess } from "@/interfaces";
-import exportPdf, { ExportPDFOptions } from "@/modules/capture/export-pdf";
-import screenshot, { CaptureScreenshotOptions } from "@/modules/capture/screenshot";
+import exportPdf, * as exportPdf_1 from "@/modules/capture/export-pdf";
+import screenshot, * as screenshot_1 from "@/modules/capture/screenshot";
 import MediaService from "@/services/MediaService";
 
 import BaseController from "./BaseController";
@@ -16,7 +16,7 @@ export default class UtilityController extends BaseController {
 	@Security("api_key")
 	@Security("jwt")
 	@Post("/export-pdf")
-	async exportWebpagePDF(@Body() body: ExportPDFOptions, @Queries() queryParams?: { url: string }) {
+	async exportWebpagePDF(@Body() body: exportPdf_1.ExportPDFOptions, @Queries() queryParams?: { url: string }) {
 		if (!this.filter.url) return respondFailure(`Param "url" is required.`);
 
 		// process
@@ -41,7 +41,7 @@ export default class UtilityController extends BaseController {
 	@Security("api_key")
 	@Security("jwt")
 	@Post("/capture-screenshot")
-	async captureScreenshot(@Body() body: CaptureScreenshotOptions, @Queries() queryParams?: { url: string }) {
+	async captureScreenshot(@Body() body: screenshot_1.CaptureScreenshotOptions, @Queries() queryParams?: { url: string }) {
 		if (!this.filter.url) return respondFailure(`Param "url" is required.`);
 
 		// process